Making sure the mountpoint exists.
If you want to allow this module to mount on any location, the mountpoint ($title) needs to exist as a directory. This addition creates the directory.
As Puppet can't do 'mkdir -p', (https://groups.google.com/forum/#!topic/puppet-users/1N0KMmurhcI) there is a limitation; the parent directories need to exist.