Use delete_undef_values() for mount options
In newer versions of Puppet, undefined variables are actually `undef`,
and not an empty string.  This causes `gluster::mount` to return a mount
option list that looks like:
```
defaults,undef,undef,undef,undef,undef
```
By using stdlib 4.2.0 or newer, we can use the `delete_undef_values()`
function.

The other alternative would be to explicitly add an `else` block to
each possible mount option's `if` to set it to the empty string.